requirements.txt
anyio==3.6.2
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
python-telegram-bot==13.15
rfc3986==1.5.0
sniffio==1.3.0

